addressing some of the code-review suggestions
